+++ src/dayjs/is-valid-duration.ts
@@ -1,0 +1,6 @@
+import type { Duration } from 'dayjs/plugin/duration'
+
+/** @hidden */
+export function isValidDuration(duration: Duration): boolean {
+    return !Number.isNaN(duration.asMilliseconds()) && duration.asMilliseconds() >= 0
+}